[#1541] removing classes and functions from file. Moved to other packages
